requirements.txt
numpy
pillow
jproperties